Refactoring!
Added an Notification listener to that updates the view to display the
weather condidtion or an error message. Swtiched the 'switch statement
out for a Nsdictionary approach. Removel the modal aspect of the code
into seperte files. in the process lost all animations, which will need
to be added back in a new fashion.